requirements.txt
azure.cognitiveservices.speech==1.22.0
boto3==1.24.10
google-cloud-texttospeech==2.11.1
gruut-ipa==0.13.0
requests==2.28.0
python-dotenv==0.20.0
lxml==4.9.0